# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LoanFlow - Loan Origination System API
  description: |
    Enterprise-grade Loan Origination System API for Indian Banks.

    ## Authentication
    All endpoints require JWT Bearer token obtained via Keycloak OAuth2.

    ## Rate Limiting
    - Standard: 100 requests/minute
    - Bulk operations: 10 requests/minute

    ## Versioning
    API version is included in the URL path (e.g., /api/v1/)
  version: 1.0.0
  contact:
    name: LoanFlow API Support
    email: api-support@loanflow.com
  license:
    name: Proprietary
    url: https://loanflow.com/license
servers:
  - url: https://api.loanflow.com/api/v1
    description: Production
  - url: https://staging-api.loanflow.com/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local Development
...
security:
  - bearerAuth: []
paths:
...
  /applications/{applicationId}:
    get:
      tags: [Applications]
      summary: Get application by ID
      operationId: getApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Applications]
      summary: Update application
      operationId: updateApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApplicationRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /applications/{applicationId}/submit:
    post:
      tags: [Applications]
      summary: Submit application for processing
      operationId: submitApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
      responses:
        '200':
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
  /applications/{applicationId}/documents/{documentId}:
    get:
      tags: [Documents]
      summary: Get document metadata
      operationId: getDocument
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
        - $ref: '#/components/parameters/DocumentIdParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetailResponse'

    delete:
      tags: [Documents]
      summary: Delete document
      operationId: deleteDocument
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
        - $ref: '#/components/parameters/DocumentIdParam'
      responses:
        '204':
          description: Deleted
...
  /applications/{applicationId}/documents/{documentId}/verify:
    post:
      tags: [Documents]
      summary: Verify document
      operationId: verifyDocument
      parameters:
        - $ref: '#/components/parameters/ApplicationIdParam'
        - $ref: '#/components/parameters/DocumentIdParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyDocumentRequest'
      responses:
        '200':
          description: Verified
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Keycloak

  # ============================================
  # PARAMETERS
  # ============================================
  parameters:
    ApplicationIdParam:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DocumentIdParam:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # SCHEMAS
  # ============================================
  schemas:
...
    ApplicationStatus:
      type: string
      enum:
        - DRAFT
        - SUBMITTED
        - DOCUMENTS_PENDING
        - UNDER_VERIFICATION
        - CREDIT_CHECK
        - POLICY_EVALUATION
        - UNDER_REVIEW
        - PENDING_APPROVAL
        - APPROVED
        - CONDITIONALLY_APPROVED
        - REJECTED
        - WITHDRAWN
        - DISBURSEMENT_PENDING
        - PARTIALLY_DISBURSED
        - DISBURSED
        - CANCELLED
    ApplicantType:
      type: string
      enum: [PRIMARY, CO_APPLICANT, GUARANTOR]
    DocumentCategory:
      type: string
      enum: [KYC, INCOME, COLLATERAL, LEGAL, OTHER]
    DocumentStatus:
      type: string
      enum: [PENDING, VERIFIED, REJECTED, EXPIRED]
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        traceId:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              rejectedValue:
                type: string
...
    UpdateApplicationRequest:
      type: object
      properties:
        requestedAmount:
          type: number
          format: double
        tenureMonths:
          type: integer
        purpose:
          type: string
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        applicationNumber:
          type: string
          example: LF20250201000123
        productCode:
          type: string
        productName:
          type: string
        requestedAmount:
          type: number
          format: double
        approvedAmount:
          type: number
          format: double
        tenureMonths:
          type: integer
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        branchCode:
          type: string
        channel:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApplicationDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApplicationResponse'
        - type: object
          properties:
            applicants:
              type: array
              items:
                $ref: '#/components/schemas/ApplicantResponse'
            documents:
              type: array
              items:
                $ref: '#/components/schemas/DocumentResponse'
            collaterals:
              type: array
              items:
                $ref: '#/components/schemas/CollateralResponse'
            offers:
              type: array
              items:
                $ref: '#/components/schemas/LoanOfferResponse'
            decision:
              type: string
            decisionDate:
              type: string
              format: date-time
            decisionRemarks:
              type: string
            riskGrade:
              type: string
...
    ApplicantResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        applicantType:
          $ref: '#/components/schemas/ApplicantType'
        fullName:
          type: string
        dateOfBirth:
          type: string
          format: date
        pan:
          type: string
        mobileNumber:
          type: string
        email:
          type: string
        monthlyIncome:
          type: number
        employmentType:
          type: string
        kycVerified:
          type: boolean
...
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        documentTypeCode:
          type: string
        documentTypeName:
          type: string
        category:
          $ref: '#/components/schemas/DocumentCategory'
        fileName:
          type: string
        fileSizeBytes:
          type: integer
          format: int64
        verificationStatus:
          $ref: '#/components/schemas/DocumentStatus'
        ocrStatus:
          type: string
        uploadedAt:
          type: string
          format: date-time
    DocumentDetailResponse:
      allOf:
        - $ref: '#/components/schemas/DocumentResponse'
        - type: object
          properties:
            extractedData:
              type: object
            confidenceScore:
              type: number
            verifications:
              type: array
              items:
                type: object
                properties:
                  verificationType:
                    type: string
                  status:
                    type: string
                  verifiedAt:
                    type: string
                    format: date-time
    VerifyDocumentRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [VERIFIED, REJECTED]
        remarks:
          type: string
...
    CollateralResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        collateralType:
          type: string
        marketValue:
          type: number
        forcedSaleValue:
          type: number
        valuationDate:
          type: string
          format: date
        status:
          type: string

    # --- UNDERWRITING ---
...
    LoanOfferResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        offerNumber:
          type: string
        offeredAmount:
          type: number
        interestRate:
          type: number
        tenureMonths:
          type: integer
        emi:
          type: number
        processingFee:
          type: number
        netDisbursement:
          type: number
        status:
          type: string
        validUntil:
          type: string
          format: date-time
        conditions:
          type: array
          items:
            type: string

    # --- COMPLIANCE ---
...